lint: [move-only] Move cpp related lints to lint_cpp.rs
